requirements.txt
aiohttp==3.9.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.7.2
asttokens==3.0.0
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
discord==2.3.2
discord.py==2.3.2
Django==4.2.6
django-ckeditor==6.7.2
django-js-asset==3.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==2.1.0
Flask==3.0.0
Flask-RESTful==0.3.10
frozenlist==1.4.0
idna==3.4
ipdb==0.13.13
ipython==8.31.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
mailtrap==2.0.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
multidict==6.0.4
oauthlib==3.2.2
openpyxl==3.1.2
parso==0.8.4
prompt_toolkit==3.0.48
psycopg2==2.9.9
pure_eval==0.2.3
pyasn1==0.5.0
Pygments==2.18.0
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.3
tweepy==4.14.0
twilio==8.10.2
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
yarl==1.9.3
